# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DisClone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/update:
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/me/status:
    patch:
      operationId: updateMyStatus
      tags: [auth]
      summary: Update presence status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
  /servers/{serverId}/channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List channels in a server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channels list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Created channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /channels/{channelId}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get channel details
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    patch:
      operationId: updateChannel
      tags: [channels]
      summary: Update channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelUpdate"
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Delete channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /channels/{channelId}/pins:
    get:
      operationId: listPinnedMessages
      tags: [channels]
      summary: List pinned messages in a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pinned messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  # ─── MESSAGES ────────────────────────────────────────────────────────────────
...
  /dms:
    get:
      operationId: listDMs
      tags: [dms]
      summary: List all DM conversations
      responses:
        "200":
          description: DM list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DMChannel"
    post:
      operationId: openDM
      tags: [dms]
      summary: Open or get a DM with a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DMInput"
      responses:
        "200":
          description: DM channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DMChannel"
  /dms/{dmId}/messages:
    get:
      operationId: listDMMessages
      tags: [dms]
      summary: List messages in a DM (latest 50)
      parameters:
        - name: dmId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: DM messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendDMMessage
      tags: [dms]
      summary: Send a DM message
      parameters:
        - name: dmId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── FRIENDS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # ─── AUTH ─────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 2
          maxLength: 32
        password:
          type: string
          minLength: 8
        displayName:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        accentColor:
          type: ["string", "null"]
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [online, idle, dnd, invisible]
        customStatus:
          type: ["string", "null"]
        customStatusEmoji:
          type: ["string", "null"]

    # ─── USER ─────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, username, discriminator, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        discriminator:
          type: string
        displayName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        accentColor:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, idle, dnd, invisible, offline]
        customStatus:
          type: ["string", "null"]
        customStatusEmoji:
          type: ["string", "null"]
        isNitro:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─── SERVER ───────────────────────────────────────────────────────────────
...
    Channel:
      type: object
      required: [id, serverId, name, type, position, createdAt]
      properties:
        id:
          type: integer
        serverId:
          type: integer
        parentId:
          type: ["integer", "null"]
        name:
          type: string
        topic:
          type: ["string", "null"]
        type:
          type: string
          enum: [text, voice, stage, forum, announcement, category]
        position:
          type: integer
        slowmode:
          type: integer
        nsfw:
          type: boolean
        userLimit:
          type: ["integer", "null"]
        bitrate:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ChannelInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [text, voice, stage, forum, announcement, category]
        topic:
          type: string
        parentId:
          type: integer
        position:
          type: integer
        slowmode:
          type: integer
    ChannelUpdate:
      type: object
      properties:
        name:
          type: string
        topic:
          type: ["string", "null"]
        position:
          type: integer
        slowmode:
          type: integer
        nsfw:
          type: boolean
        parentId:
          type: ["integer", "null"]

    # ─── MESSAGE ──────────────────────────────────────────────────────────────
    Message:
      type: object
      required: [id, channelId, authorId, content, createdAt]
      properties:
        id:
          type: integer
        channelId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
        editedAt:
          type: ["string", "null"]
        isPinned:
          type: boolean
        replyToId:
          type: ["integer", "null"]
        replyTo:
          $ref: "#/components/schemas/MessageSnippet"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
        embeds:
          type: array
          items:
            $ref: "#/components/schemas/Embed"
        mentions:
          type: array
          items:
            type: integer
        mentionsEveryone:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MessageSnippet:
      type: object
      required: [id, authorId, content]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000
        replyToId:
          type: integer
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentInput"
        mentions:
          type: array
          items:
            type: integer
        mentionsEveryone:
          type: boolean
...
    Attachment:
      type: object
      required: [id, url, filename, contentType, size]
      properties:
        id:
          type: integer
        url:
          type: string
        filename:
          type: string
        contentType:
          type: string
        size:
          type: integer
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
    AttachmentInput:
      type: object
      required: [url, filename, contentType, size]
      properties:
        url:
          type: string
        filename:
          type: string
        contentType:
          type: string
        size:
          type: integer
        width:
          type: integer
        height:
          type: integer
    Reaction:
      type: object
      required: [emoji, count, userIds]
      properties:
        emoji:
          type: string
        count:
          type: integer
        userIds:
          type: array
          items:
            type: integer
        me:
          type: boolean
    Embed:
      type: object
      properties:
        type:
          type: string
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        color:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        authorName:
          type: ["string", "null"]
        footerText:
          type: ["string", "null"]

    # ─── MEMBER ───────────────────────────────────────────────────────────────
...
    DMChannel:
      type: object
      required: [id, type, participants, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [dm, group_dm]
        name:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        lastMessageAt:
          type: ["string", "null"]
        lastMessage:
          $ref: "#/components/schemas/MessageSnippet"
        createdAt:
          type: string
          format: date-time
    DMInput:
      type: object
      required: [recipientId]
      properties:
        recipientId:
          type: integer

    # ─── FRIEND ───────────────────────────────────────────────────────────────
...